+++ src/components/functional/toasts-list/components/transaction-toast/transaction-toast.tsx
@@ -1,6 +1,6 @@
 import type { EventEmitter, JSX } from '@stencil/core';
 import { Component, Event, h, Prop } from '@stencil/core';
-import type { ITransactionListItem } from 'components/visual/transaction-list-item/transaction-list-item.types';
+import type { ITransactionListItem } from 'components/visual/TransactionListItem/transactionListItem.types';
 import { TransactionStatusEnum } from 'constants/transactionStatus.enum';
 
 import type { IToastDataState, ITransactionProgressState } from './transaction-toast.type';