+++ src/components/visual/TransactionListItem/transactionListItem.styles.ts
@@ -1,0 +1,24 @@
+// prettier-ignore
+export default {
+    transactionItem: 'transaction-item mvx:flex mvx:items-center mvx:py-5 mvx:px-3 mvx:gap-3 mvx:rounded-lg mvx:transition-all mvx:duration-200 mvx:ease-in-out mvx:bg-secondary mvx:hover:bg-hover mvx:xxs:gap-4 mvx:xxs:px-4',
+    transactionIcon: 'transaction-icon mvx:w-8 mvx:h-8 mvx:rounded-full mvx:text-primary',
+    transactionIconFailed: 'transaction-icon-failed mvx:text-error',
+    transactionIconImg: 'transaction-icon-img mvx:w-full mvx:h-full mvx:object-cover',
+    transactionIconText: 'transaction-icon-text mvx:text-xl mvx:w-10 mvx:h-10 mvx:text-primary',
+    transactionIconSvg: 'transaction-icon-svg mvx:w-8 mvx:h-8 mvx:text-primary',
+    transactionDetails: 'transaction-details mvx:flex-1 mvx:min-w-0',
+    transactionDetailsHeader: 'transaction-details-header mvx:flex mvx:items-center mvx:justify-between',
+    transactionHeaderLeft: 'transaction-header-left mvx:flex mvx:gap-1 mvx:items-center mvx:justify-center mvx:fill-link',
+    transactionTitle: 'transaction-title mvx:text-xs mvx:font-medium mvx:leading-6 mvx:tracking-tight mvx:m-0 mvx:text-tertiary-text mvx:xxs:text-sm mvx:xs:text-base',
+    transactionTitleTruncate: 'transaction-title-truncate mvx:truncate mvx:max-w-30 mvx:shrink-0',
+    transactionDetailsInfo: 'transaction-details-info mvx:flex mvx:items-center mvx:gap-1 mvx:self-stretch mvx:text-sm mvx:font-medium mvx:leading-4 mvx:tracking-tight mvx:text-label',
+    transactionDetailsInfoText: 'transaction-details-info-text mvx:text-sm mvx:font-medium mvx:leading-4 mvx:tracking-tight mvx:max-w-64 mvx:text-label',
+    transactionDetailsInfoIcon: 'transaction-details-info-icon mvx:rounded-full mvx:w-4 mvx:h-4',
+    transactionDetailsInfoIconImg: 'transaction-details-info-icon-img mvx:rounded-full mvx:w-full mvx:h-full',
+    transactionAmount: 'transaction-amount mvx:flex mvx:text-xs mvx:font-medium mvx:leading-6 mvx:tracking-tight mvx:text-tertiary-text mvx:xxs:text-sm mvx:xs:text-base',
+    transactionAmountPositive: 'transaction-amount-positive mvx:text-success',
+    transactionAmountNegative: 'transaction-amount-negative',
+    transactionFailed: 'transaction-failed mvx:text-error',
+    transactionAmountDecimal: 'transaction-amount-decimal mvx:opacity-70',
+    transactionAmountLabel: 'transaction-amount-label mvx:opacity-70 mvx:tracking-tight mvx:truncate mvx:max-w-30 mvx:shrink-0'
+} satisfies Record<string, string>;